+++ brightroute/src/main/java/com/brightroute/brightroute/repository/LecturePartRepository.java
@@ -3,4 +3,4 @@ package com.brightroute.brightroute.repository;
 import com.brightroute.brightroute.model.LecturePart;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LecturePartRepository extends JpaRepository<LecturePart, Long> {}
+public interface LecturePartRepository extends JpaRepository<LecturePart, Integer> {}
